# Add gather_backward op