--- a/.gitignore
+++ b/.gitignore
@@ -7,6 +7,7 @@ bin-release/
 # Other files and folders
 .settings/
 __pycache__/
+.vscode/
 
 # Distribution files
 build/