--- a/tests/test_reduction_ops.py
+++ b/tests/test_reduction_ops.py
@@ -508,7 +508,9 @@ def test_accuracy_scatter_mul(src_shape, inp_shape, dim, dtype):
 @pytest.mark.parametrize("dim", [0, 1, 2])
 @pytest.mark.parametrize("dtype", FLOAT_DTYPES)
 def test_accuracy_gather(inp_shape, dim, dtype):
-    inp = torch.randn(inp_shape, dtype=dtype, device=flag_gems.device)
+    inp = torch.randn(
+        inp_shape, dtype=dtype, device=flag_gems.device, requires_grad=True
+    )
     size_dim = inp_shape[dim]
 
     import random
@@ -539,6 +541,15 @@ def test_accuracy_gather(inp_shape, dim, dtype):
         res_out = torch.gather(inp, dim, index)
 
     gems_assert_equal(res_out, ref_out)
+
+    out_grad = torch.randn_like(res_out)
+    ref_grad = to_reference(out_grad)
+
+    (ref_in_grad,) = torch.autograd.grad(ref_out, ref_inp, ref_grad)
+    with flag_gems.use_gems():
+        (res_in_grad,) = torch.autograd.grad(res_out, inp, out_grad)
+    res_in_grad = to_reference(res_in_grad)
+    gems_assert_equal(res_in_grad, ref_in_grad)
 
 
 @pytest.mark.select_scatter