@0x45f 2024-12-17
Why is the speedup so low?