@0x45f 2024-12-19
> It seems that the speedup ratio is not very ideal

I did some profiling work and found some code that can be optimized. I will do some tests locally first.