@0x45f 2024-12-24
i modify gather backwar:
1. Using new_zeros instead of zeros_like 
2. Using inplace scatter

before
```
Operator: gather_backward backward  Performance Test (dtype=torch.float16, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.021504            0.134144               0.160          [torch.Size([64, 64]), 0, torch.Size([9, 33])]
SUCCESS               0.021504            0.126976               0.169          [torch.Size([256, 256]), 0, torch.Size([254, 231])]
SUCCESS               0.027648            0.135168               0.205          [torch.Size([1024, 1024]), 1, torch.Size([778, 430])]
SUCCESS               0.144384            0.219136               0.659          [torch.Size([4096, 4096]), 0, torch.Size([3997, 233])]
SUCCESS               0.443392            0.652288               0.680          [torch.Size([1024, 65536]), 1, torch.Size([887, 277])]
SUCCESS               0.021504            0.136192               0.158          [torch.Size([1024, 256]), 1, torch.Size([546, 118])]
SUCCESS               0.040960            0.120832               0.339          [torch.Size([1024, 4096]), 0, torch.Size([651, 251])]


Operator: gather_backward backward  Performance Test (dtype=torch.float32, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.014336            0.133120               0.108          [torch.Size([64, 64]), 0, torch.Size([4, 2])]
SUCCESS               0.019456            0.132096               0.147          [torch.Size([256, 256]), 1, torch.Size([111, 217])]
SUCCESS               0.029696            0.132096               0.225          [torch.Size([1024, 1024]), 0, torch.Size([455, 897])]
SUCCESS               0.343040            1.383424               0.248          [torch.Size([4096, 4096]), 1, torch.Size([1910, 2832])]
SUCCESS               5.706752            6.578176               0.868          [torch.Size([1024, 65536]), 0, torch.Size([449, 60242])]
SUCCESS               0.019456            0.130048               0.150          [torch.Size([1024, 256]), 1, torch.Size([45, 214])]
SUCCESS               0.072704            0.167936               0.433          [torch.Size([1024, 4096]), 0, torch.Size([381, 2429])]


Operator: gather_backward backward  Performance Test (dtype=torch.bfloat16, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.032768            0.130048               0.252          [torch.Size([64, 64]), 0, torch.Size([43, 55])]
SUCCESS               0.029696            0.132096               0.225          [torch.Size([256, 256]), 0, torch.Size([156, 146])]
SUCCESS               0.033792            0.132096               0.256          [torch.Size([1024, 1024]), 1, torch.Size([806, 71])]
SUCCESS               0.649216            1.077248               0.603          [torch.Size([4096, 4096]), 1, torch.Size([1989, 3312])]
SUCCESS               1.816576            4.820992               0.377          [torch.Size([1024, 65536]), 1, torch.Size([355, 48120])]
SUCCESS               0.036864            0.125952               0.293          [torch.Size([1024, 256]), 1, torch.Size([178, 225])]
SUCCESS               0.097280            0.138240               0.704          [torch.Size([1024, 4096]), 0, torch.Size([336, 3222])]
```

after:
```
Operator: gather_backward backward  Performance Test (dtype=torch.float16, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.022528            0.056320               0.400          [torch.Size([64, 64]), 0, torch.Size([9, 33])]
SUCCESS               0.021504            0.021504               1.000          [torch.Size([256, 256]), 0, torch.Size([254, 231])]
SUCCESS               0.027648            0.046080               0.600          [torch.Size([1024, 1024]), 1, torch.Size([778, 430])]
SUCCESS               0.144384            0.199680               0.723          [torch.Size([4096, 4096]), 0, torch.Size([3997, 233])]
SUCCESS               0.443392            0.447488               0.991          [torch.Size([1024, 65536]), 1, torch.Size([887, 277])]
SUCCESS               0.021504            0.025600               0.840          [torch.Size([1024, 256]), 1, torch.Size([546, 118])]
SUCCESS               0.038912            0.046080               0.844          [torch.Size([1024, 4096]), 0, torch.Size([651, 251])]


Operator: gather_backward backward  Performance Test (dtype=torch.float32, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.013312            0.015360               0.867          [torch.Size([64, 64]), 0, torch.Size([4, 2])]
SUCCESS               0.017408            0.017408               1.000          [torch.Size([256, 256]), 1, torch.Size([111, 217])]
SUCCESS               0.029696            0.052224               0.569          [torch.Size([1024, 1024]), 0, torch.Size([455, 897])]
SUCCESS               0.344064            0.861184               0.400          [torch.Size([4096, 4096]), 1, torch.Size([1910, 2832])]
SUCCESS               5.700608            5.012480               1.137          [torch.Size([1024, 65536]), 0, torch.Size([449, 60242])]
SUCCESS               0.019456            0.018432               1.056          [torch.Size([1024, 256]), 1, torch.Size([45, 214])]
SUCCESS               0.072704            0.122880               0.592          [torch.Size([1024, 4096]), 0, torch.Size([381, 2429])]


Operator: gather_backward backward  Performance Test (dtype=torch.bfloat16, mode=cuda, level=comprehensive)
Size         Torch Latency (ms)    Gems Latency (ms)         Gems Speedup         Size Detail
------------------------------------------------------------------------------------------
SUCCESS               0.031744            0.014336               2.214          [torch.Size([64, 64]), 0, torch.Size([43, 55])]
SUCCESS               0.029696            0.017408               1.706          [torch.Size([256, 256]), 0, torch.Size([156, 146])]
SUCCESS               0.032768            0.024576               1.333          [torch.Size([1024, 1024]), 1, torch.Size([806, 71])]
SUCCESS               0.649216            0.613376               1.058          [torch.Size([4096, 4096]), 1, torch.Size([1989, 3312])]
SUCCESS               1.816576            3.298304               0.551          [torch.Size([1024, 65536]), 1, torch.Size([355, 48120])]
SUCCESS               0.035840            0.020480               1.750          [torch.Size([1024, 256]), 1, torch.Size([178, 225])]
SUCCESS               0.096256            0.102400               0.940          [torch.Size([1024, 4096]), 0, torch.Size([336, 3222])]
```