@tongxin 2025-01-02
This PR was further improved in [PR 382](https://github.com/FlagOpen/FlagGems/pull/382), along with scatter_add/mul reimplemented using tl atomics. Let's discontinue this PR and thanks for your contribution!